add equalsTest for GameDbData with Mockito
