+++ app/build.gradle
@@ -89,6 +89,9 @@ dependencies {
 
     implementation "com.google.android.gms:play-services-location:18.0.0"
 
+    //Mockito
+    testCompile 'org.mockito:mockito-core:3.7.7'
+
 }
 
 tasks.withType(Test) {
