+++ app/src/androidTest/java/sdp/moneyrun/MenuActivityTest.java
@@ -64,6 +64,11 @@ public class MenuActivityTest {
     }
 
     @Test
+    public void activityStartsProperly() {
+        assertEquals(State.RESUMED, testRule.getScenario().getState());
+    }
+
+    @Test
     public void joinGamePopupIsDisplayed() {
         try (ActivityScenario<MenuActivity> scenario = ActivityScenario.launch(MenuActivity.class)) {
             Intents.init();
