+++ app/src/main/java/sdp/moneyrun/MenuActivity.java
@@ -224,64 +224,6 @@ public class MenuActivity extends AppCompatActivity implements NavigationView.On
     }
 
 
-    public void onButtonShowQuestionPopupWindowClick(View view, Boolean focusable, int layoutId, Riddle riddle) {
-
-        PopupWindow popupWindow = onButtonShowPopupWindowClick(view, focusable, layoutId);
-        TextView tv = popupWindow.getContentView().findViewById(R.id.question);
-        int correctId = 0;
-
-        //changes the text to the current question
-        tv.setText(riddle.getQuestion());
-
-        int[] buttonIds = {R.id.question_choice_1, R.id.question_choice_2, R.id.question_choice_3, R.id.question_choice_4};
-        TextView buttonView = tv;
-
-        //Loops to find the ID of the button solution and assigns the text to each button
-        /*for (int i = 0; i < 4; i++){
-
-            buttonView = popupWindow.getContentView().findViewById(buttonIds[i]);
-            buttonView.setText(riddle.getPossibleAnswers()[i]);
-
-            if(riddle.getPossibleAnswers()[i].equals(riddle.getAnswer()))
-                correctId = buttonIds[i];
-        }*/
-
-        popupWindow.getContentView().findViewById(correctId).setOnClickListener(new View.OnClickListener() {
-
-            @Override
-            public void onClick(View v) {
-                popupWindow.dismiss();
-            }
-        });
-
-    }
-
-    /**
-     *
-     * @param view Current view before click
-     * @param focusable Whether it can be dismissed by clicking outside the popup window
-     * @param layoutId Id of the popup layout that will be used
-     */
-    public PopupWindow onButtonShowPopupWindowClick(View view, Boolean focusable, int layoutId) {
-
-        // inflate the layout of the popup window
-        LayoutInflater inflater = (LayoutInflater)
-                getSystemService(LAYOUT_INFLATER_SERVICE);
-        View popupView = inflater.inflate(layoutId, null);
-
-        // create the popup window
-        int width = LinearLayout.LayoutParams.WRAP_CONTENT;
-        int height = LinearLayout.LayoutParams.WRAP_CONTENT;
-        final PopupWindow popupWindow = new PopupWindow(popupView, width, height, focusable);
-
-        // show the popup window at wanted location
-        popupWindow.showAtLocation(view, Gravity.CENTER, 0, 0);
-
-        return popupWindow;
-    }
-
-
-
 //    @Override
 //    public boolean onNavigationItemSelected(@NonNull MenuItem item) {
 //        switch (item.getItemId()){
